+++ lib/src/format.dart
@@ -53,3 +53,16 @@ String fileExt(String filePath) {
 
   return pSpl.isNotEmpty ? pSpl.last : 'Unknown';
 }
+
+/// Any argument value will be attempted to be parsed as a Number type without returning an error. If parsing fails, it is replaced by the number set in `fallback`. The default value for `fallback` is `0`. You can specify `radix` (default is decimal: `10`) in the third argument.
+int safeParseInt(dynamic value, {int? fallback, int? radix}) {
+  if (value == null) {
+    return fallback ?? 0;
+  }
+
+  try {
+    return int.parse(value.toString().split('.').first, radix: radix ?? 10);
+  } catch (e) {
+    return fallback ?? 0;
+  }
+}