+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `fileSize` method
 - Add `fileExt` method
+- Add `safeParseInt` method
 
 ## 0.0.4 (2024-10-02)